Move all cc source from backends to codegen
